removed code duplication in chan2detect by moving make_masks() to masks.py
